Major refactoring and some visible changes. The source code was already a mess, and any extension would have likely suffered from that. Explicit model "classes" (strictly speaking, javascript has no classes) allow for a clear separation between model and view. The use of client-side templating also makes the code easier to understand. The application uses the Nunjucks templating system, which is Mozilla's javascript version of the elegant Jinja2 templating system for Python. The control code is still scattered throughout the app, and not all program parts conform to the new architecture, so more refactoring will have to be done. Refactoring took place mainly in the front-end, but the backend has also been improved. In particular, there is now a clear separation between core classes and the networking part (URL dispatching, JSON serialization/deserialization).
The editing of conceptual scales and context families previously required the user to double-click on table rows, which might be
perceived as somewhat tedious. Moreover, each double-click event additionally resulted in two executions of the single-click
event handler, which (besides being ugly) eventually resulted in erroneous behavior. So the app no longer uses double-clicks.
Editing (and a few other options) are now realized by explicitly labeled buttons rather than cryptic buttons in a subwindow's titlebar
(which could not take up too much space because of the title). When the user enters a new database (URL and DB name), its tables,
columns and their SQL types are shown to the user. For each sort (i.e. table), there is now an explicity stored SQL expression detailing
the string representation of that sort's objects (i.e. tuples).

The layout is now more consequently based on the Flexbox model. 2-D scrollable tables are now supported, but the solution is somewhat
complicated and not yet 100% appealing (essentially because the browser allows only limited control over its scrollbars, so custom
scrollbars look like the better option). A dropdown menu allows the user to select between different context families to navigate in,
but there seems to be a bug in this version. The options pane has been visually improved but is still not appealing.

This version of the app was presented at ICFCA 2017 (poster session).