new data backend "FacetedPowerContextFamily"
A "faceted power context family" is essentially a power context family (as defined
in Formal Concept Analysis), but is allowed to contain several relational contexts
(called "facets") for each arity. The attributes of its single object context are
considered to be sorts. Each n-ary facet has an associated tuple (s_1,...,s_n) of
sorts, which limits the facet's (formal) objects to tuples (g_1,...,g_n) of objects
of the respective sorts. A facet's (formal) attributes are thought of as semantically
related (reflecting the general understanding of "facet"). A power context family
(in the usual sense) is obtained by assembling all facets of the same arity into
a single relation context. Working with facets (instead of a single relation context)
better supports the requirements of navigation, as implemented by this package.

This commit provides only the basic Python classes, JSON encoding/decoding and
notably query answering (based on homomorphism search). The new data backend
is not yet integrated in the application.