refactoring (preparing for extensions)
changes in this and the previous commit:
There is now a more abstract and well-defined interface between DBContextFamily and the surrounding code, to enable switching
data backends. To that end, an explicit context class "ScaledContext" was introduced and is exposed by DBContextFamily
(previously, many-valued attributes and scaled contexts were conflated, which was confusing and made surrounding code
rely on DBContextFamily internals).

The first implementation of ScaledContext consisted of a many-valued attribute and a conceptual scale, which seemed natural from
a conceptual point of view. However, the following problem was encountered: If the scaled context relies on an interval scale,
it should provide a boundary interval of available values in the database. Because ScaledContext is generic, all interval logic
is encoded in the scale, so the scale must provide suitable SQL fragments for querying the boundary interval. The query itself
is executed by ScaledContext (the scale shouldn't know the database), but the scale is then again needed for interpreting
the query results. So it seems that a rather complex interaction between ScaledContext and scale would be required to compute
the boundary interval.

The chosen solution: the scale is NOT realized as a member of ScaledContext. Instead, ScaledContext is realized as an abstract
base class, and is superclassed to provide the desired scale's functionality (the scale provides just functions/behavior,
whereas the many-valued attribute provides the instance attributes). When we think of a ScaledContext as a pattern context,
the scale takes the role of the pattern space; from this angle, it seems indeed natural to represent the "scale" by a class.