bugfix
It is possible to change both the sort (using checkboxes) and the display attributes
(using the eye icon) on a graph node. But when changing the sort, the selected display
attributes are obviously invalid for the new sort. This has resulted in errors.
Now, when changing the sort, all selected display attributes are discarded.