DB autoconf: can read composite primary keys
So far, the program could not automatically create an initial binding
(i.e. navigation config) for a database if one of its tables has
a composite primary key. This is now possible; the print function for
such tables concatenates all primary key columns. With this feature,
the Northwind database can be read in. An autogenerated JSON binding
is included in this commit.