Major refactoring: Hierarchical MVC
The controller class was getting too large and unwieldy, so instead there are now several controllers
(in particular for the sub-applications: navigation, bindings editor). These are organized in a hierarchy,
with a main/root controller at the top. A new class "MasterControl" encapsulates the hierarchy and provides
a single point of access. A new variable "ctrl" gets submitted with each HTTP request, so that we know which
controller the front end is talking to.