minor refactoring: HTTP request handling
The signature of HTTP request handlers is changed. They now receive a HttpRequest (which is added as a new class)
and return a HttpResponse (as before). This allows request handlers to access HTTP header information
(which was not the case previously). Request handling is modeled after the Django frameword (but is much simpler, and done in
the WSGI file itself).