New feature: HTTP Sessions
Application state was previously sent by the client, along with the request,
and essentially decoded on the server. It is now stored as a Python pickle in a file
/tmp/granada_session_<session_id> (or whatever is the system-wide temp directory
on the server). This is not necessarily a permanent change, but it gets application
state out of the way for some upcoming major refactoring.